Extremely simple model base
In this version, cars do not move and don't exhibit any behvaiour. A results class was implemented such that front end can easily grab the results of the simulation. In this simulation, all four junctions are created and the traffic control object signals green and red to switch between them and is green for the specific green time indicated by the user. Pedestrian crossings are not implemented yet though I dont think will be difficult. Simpy integration has been successful, all that is needed is for additional features to be built on top of this base